Thread Safety
exceptions, tralloc, and randoming are now thread safe (per file) by using mutexing.h internally
thread safety can be disabled for exceptions and tralloc with TRALLOC/EXCEPTIONS_NOTHREADSAFE,
but it is builtin to randoming (manually in RANDOMING_USE_RAND option since rand() is not thread safe) to conform with winapi and unix random functions, which are thread safe.
The default randoming implementation on windows requires -lbcrypt (cryptographically secure prng) on command line, which is available from vista onwards.
On unix, it reads either from urandom or random, if urandom cannot be opened
If bcrypt is not available or OS is not windows or unix, RANDOMING_USE_RAND option can be used to rely on the c stdlib but is likely to be inferior and vary in randomness per implementation.
On windows, rand implementation uses time(NULL) * GetCurrentProcessId() to seed rand.
On unix,  time(NULL) * getpid() is used to seed rand
If neither windows or unix, but C11 threads are supported, time(NULL) * thrd_current() is used to seed rand
If neither win or unix, and threads are not supported, allocation (and freeing) of one byte is abused to get a varying address of data, then time(NULL) * (uintptr_t)addr is used to seed rand. This is likely the worst case scenario.
Added mutex_create_constructor, a macro that accepts a _Bool and a mutex, allows initializing a mutex in a constructor without exit potentially being called and invoking undefined behavior.
Also added mutex_delete_destructor, a macro that accepts a _Bool and a mutex, allows destroying a mutex in a destructor without exit potentially being called and invoking undefined behavior.